[WIP] pkg/gui: Only suggest remote branches when creating a PR against a selected branch
When creating a PR against a selected branch (via O = "create pull request
options"), the suggestion area is now populated with all remote branches -
instead of all local ones. After all, creating a PR against a branch that
doesn't exist on the remote won't work.

To support "PR is not filed against 'origin' remote" use cases (e.g. when
contributing via a fork that is 'origin' to a GitHub project that is
'upstream'), branches of all remotes (and not only 'origin') are listed. This
also simplifies the implementation.

Fixes #1826.